Add Candidates
The file which contains the code responsible for receiving and storing information of the candidates standing for the election. Since I did this file for the school head boy and head girl election, and my school only allowed grade 9 and 11 students to stand in the elections, the corresponding restriction is put in the code.