Add files via upload
Contains the code responsible for receiving and storing the information of candidates eligible to stand for the election. Since I created this program for my school head boy and head girl election, and sinc my school only allowed grade 9 amd 11 students to compete in the elections, the corresponding restriction is in the code.