Add files via upload
Contains the code responsible for receiving and storing the information of the voters who are going to vote for the election. Since I made this code for my school head boy and head girl elections, the file receives and stores the name and ID number of the students voting in the election.